career-assistant/requirements.txt
gradio==4.44.1
gradio-client==1.3.0
pydantic==2.10.6

scikit-learn==1.5.1
numpy==2.1.1
pandas==2.2.2
PyYAML==6.0.2
joblib==1.4.2
requests==2.32.3
python-dotenv==1.0.1
pypdf2==3.0.1
nltk==3.9.1
